Changed weapon power to float and changed prefabs to use different damage to make laser less OP.

+++ Assets/Scripts/Weapon.cs
@@ -3,7 +3,7 @@
 public class Weapon : MonoBehaviour
 {
 	[Header("Weapon: Inspector Set Fields")]
-	public int damagePower = 1;
+	public float damagePower = 1;
 	public float coolDownTime = 0.5f;
 	public float maxDistance;
